Don't render tab panels unless they are clicked
cf-component-tabs would render all the tabpanel components when a tab
component would get rendered. This has the potential of loading a bunch
of data over the network that the user will never see.

By changing it this way we are able to only request the data needed to
render the tabs the user actually clicks on