card: Keep a local state of drawers if no props
Whenever you're using the `CardDrawers` sub-component of `Card` you need
to add some keys to your state and handle all the events yourself. This
is fine if you want to control the flow very tightly but the most common
case is just to switch to the drawers selected.

This patch changes the `CardDrawers` sub-component so it still takes in
the props and those will control the component but falls back to the
local state if the props `active` and `onClick` are not present.