Select: Pass the searchable prop to react-select (#167)
* Select: Pass the searchable prop to react-select

Since we use the rest syntax when deconstructing the props we do not have the props we pick out (`async`, `creatable`, `searchable`, `label`) in the `props`. This means that when we spread the object over the `react-select` component it will not include those values. `react-select` defaults the `searchable` prop to `true` which is what we explicitly don't want and set to `false` in the `defaultProps`.

This patch explicitly passes the `searchable` prop down to the `react-select` component.

* Select: Update jest snapshot

* Notifications: Fix lint error

Fixed by running `eslint --fix`