Sync Legacy into Master (#168)
* Fix JSON content-type detection in 'wrapResponse' (#140)

* #139 Fix JSON content-type detection in 'wrapResponse'

* add test for json content type in the response and add credentials to same origin by default

* Publish

 - cf-util-http-file@4.5.3
 - cf-util-http-poll@2.5.3
 - cf-util-http@3.5.3

* card: Keep a local state of drawers if no props (#125)

* card: Keep a local state of drawers if no props

Whenever you're using the `CardDrawers` sub-component of `Card` you need
to add some keys to your state and handle all the events yourself. This
is fine if you want to control the flow very tightly but the most common
case is just to switch to the drawers selected.

This patch changes the `CardDrawers` sub-component so it still takes in
the props and those will control the component but falls back to the
local state if the props `active` and `onClick` are not present.

* card: Explicitly check for active prop in drawsers

Since the state property can have the very reasonable value `null` which
signifies no drawer should be shown we need to make a explicit
`hasOwnProperty` call to make sure that the `active` prop has been set
and isn't just falsy because it wasn't set.

* Publish

 - cf-builder-card@2.3.0
 - cf-component-card@1.4.0

* Publish

 - cf-builder-table@2.2.1
 - cf-component-button@5.0.0
 - cf-component-heading@4.0.0
 - cf-component-modal@6.1.0
 - cf-component-text@4.0.0

* avoid use of Headers.entries() and for .. of headers because of limited support in Edge browser
more details: https://developer.mozilla.org/en-US/docs/Web/API/Headers\#Browser_compatibility

* add comment about use of Map

* Publish

 - cf-util-http-file@4.5.4
 - cf-util-http-poll@2.5.4
 - cf-util-http@3.5.4

* fix flaky test with removal of Date.now() checks

* reverted changes to logic with Headers for Edge, polyfill fetch for Edge instead

* use --skip-temp-tag for publishing

* Publish

 - cf-util-http-file@4.5.5
 - cf-util-http-poll@2.5.5
 - cf-util-http@3.5.5

* Select: Pass the searchable prop to react-select (#167)

* Select: Pass the searchable prop to react-select

Since we use the rest syntax when deconstructing the props we do not have the props we pick out (`async`, `creatable`, `searchable`, `label`) in the `props`. This means that when we spread the object over the `react-select` component it will not include those values. `react-select` defaults the `searchable` prop to `true` which is what we explicitly don't want and set to `false` in the `defaultProps`.

This patch explicitly passes the `searchable` prop down to the `react-select` component.

* Select: Update jest snapshot

* Notifications: Fix lint error

Fixed by running `eslint --fix`

* Card: Pass the id of the drawer clicked to parent (#160)

When the local state was added to the `CardDrawers` component we failed
at passing the prop back up to the parent causing the drawers not
opening when the links are clicked.

This patch passes the id up to the `onClick` function when a link is
clicked.

Fixes #159