icon: set types directly in proptypes
Instead of keeping the types and sizes available to the `Icon` component
in a different file and injecting it into the prop types we now do it
explicitly in the component proptypes.

This means that the documentation that's generated will show the types
available instead of just `TYPES` and `SIZES`.

Closes: #192