icon: remove throw in default case
If a `type` we do not recognize is passed down as a prop to the `Icon`
component we used to throw an error. Instead we should trust that
developers follow the restrictions from prop types and supply a valid
`type`.

This patch removes the throwing of a error when a `type` property isn't
recognized and the results of doing it are un-documented dev error.