chore(tooling): force conventional commits
Add the packages `husky` and `validate-commit-msg` to create a precommit
git hook that rejects commits that do not adhere to the conventional
commits standard[1].

This enables us to run `lerna publish` with the flag
`--conventional-commits` which will autogenerate a changelog and set the
package versions automatically. This PR enables this flag.

[1]: https://conventionalcommits.org/

Closes: #237