fix(cf-component-progress): media query fix
Media queries in `cf-ui` are reversed to the ones in the legacy CSS
styles resulting in mobile styles being applied when the browser width
is larger than a average mobile device and the desktop styles being
applied when smaller than the average mobile device.

This PR reversed the styles along the media query "axis". So the correct
device styles are applies for the Progress component.