fix(cf-style-container): Change fela wrapper names (#300)
Change the component display name for fela higher order components so
that the actual components can be targeted by their display names in
tests or otherwise.

BREAKING CHANGE: If you are targeting components in your tests you are
going to need to make some changes. Previously others have selected the
second one in the node list after querying for the display name.