fix(cf-style-container): set themed component name
Sets the `displayName` property of themed components to
`Themed[Component name]` rather than the `ThemedComponent` it is set to
now.

This change will make it possible to select these components in
`enzyme` tests by their display names.

BREAKING CHANGE: The display name of themed components now have a
`Themed` prefix to their display names. So if you had a themed component
that has the name `Button` it will now be `ThemedButton` instead of
`ThemedComponent`.

Related-to: #300