Migate tabs to fela (#314)
* feat(cf-component-tabs): rewrite styling in fela

BREAKING CHANGE: This change is a significant change in the component
internals warranting a major bump in version number.

Closes: #209

* fix(cf-component-tabs): encapsulate content

Previously if a tabs panel content has had some top margin it would push
the tabs and the content apart resulting in a break in the border where
the content and the tabs meet. This was due to the content and the tabs
having separate borders made to look like the same border by being
pushed up to each other.

This patch sets the border on the whole component so it encapsulates all
the content and makes changes to the tabs and content to accommodate
this.