fix(cf-component-tabs): set the correct colors (#318)
In #314 we errounisly made assumptions where colors are kept in the
theme. Trying to access the white color at `theme.color.white` and the
smoke color at `theme.color.smoke` resulted in `undefined` values that
`fela` swallows up and not emitting a rule.

This patch references the colors in the correct location as well as
minor housekeeping to keep the tabs theme a bit tidier.