chore: add default code reviewers
Adding the code reviewers is a tedious part of opening up a PR in cf-ui.

This patch adds users as "Code Owners" of all the files in the repo so
that they get added as default reviewers for every PR that's opened.

See https://github.com/blog/2392-introducing-code-owners for more info.