feat(cf-component-tabs): `TabsItem` own component
In order to be able to apply styles to a "tab item" we need to abstract
it out into it's own component. That can be optionally be passed down to
the `Tabs` component defaulting to the styles previously applied.

This change is backwards compatible. See
`packages/cf-component-tabs/example/basic/component.js` for the optional
component property.