feat(cf-component-tabs): selected color override (#328)
Even tho the current theme doesn't make any distinction on colors when a
item is selected or not there are instances in the dashboard where we'd
like to do exactly that. Unfortunately due to the styling model we have
now we need to specifically list out any possible rules we might use
when the component is implemented, warranting this change.